Added actual project files
Added files:
index.php => Actual PHP file for production. Enter the DB credentials in this file and upload it anywhere to start using it. You can adjust settings in the $settings array and you can add or edit translations in the $language array.
index.php.info => Copy of the PHP file with comments. Remove the .info extension and open the file as PHP in your favorite IDE. You can view comments describing every part of the code that is functional within this project.